Structure
- Implementation of a common structure between screens.
