+++ BrainMe/app/sign-up.tsx
@@ -1,14 +1,16 @@
-import { View, Text, ScrollView, StyleSheet } from "react-native";
+import { Text } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { useRouter } from "expo-router";
 import { useUser, useSignUp } from "@clerk/clerk-expo";
 import React, { useEffect, useState } from "react";
 
+// This imports the structure component, which contains the main container for the screen with the header.
+import { Structure } from "@/components/auth/structure";
+
 // These imports are required to use the components in this file.
-import Header from "@/components/auth/header";
 import Input from "@/components/auth/input";
-import Footer from "@/components/auth/footer";
-import Button from "@/components/button";
+import Footer from "@/components/auth/footer-text";
+import Button from "@/components/auth/action-button";
 
 export default function Welcome() {
   // This hook provides the safe area insets, which allows you to avoid the status bar.
@@ -77,66 +79,47 @@ export default function Welcome() {
   }, [isLoaded, isSignedIn]);
 
   return (
-    <View style={{ flex: 1, paddingTop: insets.top }}>
-      <Header title="Sign up now" subtitle="Join BrainMe and invite friends" />
-      <ScrollView
-        automaticallyAdjustKeyboardInsets={true}
-        showsVerticalScrollIndicator={false}
-        style={styles.container}
-        contentContainerStyle={{ gap: 34 }}
-      >
-        {!pendingVerification ? (
-          <>
-            <Input
-              title="Email"
-              placeholder="winner@email.com"
-              keyboardType="email-address"
-              onChangeText={setEmailAddress}
-            />
-            <Input
-              title="Your username"
-              placeholder="Rookie"
-              onChangeText={setUsername}
-            />
-            <Input
-              title="Password"
-              placeholder="Insert password..."
-              secureTextEntry
-              onChangeText={setPassword}
-            />
-            <Button text="SIGN UP" onPress={onSignUpPress} />
-            <Footer text="Already have an account?" />
-          </>
-        ) : (
-          <>
-            <Text>
-              Please, insert the verification code we provided on your email
-              address.
-            </Text>
-            <Input
-              title="Verification code"
-              placeholder="Insert code..."
-              keyboardType="numeric"
-              textAlign="center"
-              maxlength={6}
-              onChangeText={setCode}
-            />
-            <Button text="VERIFY" onPress={onPressVerify} />
-            <Footer text="Already have an account?" />
-          </>
-        )}
-      </ScrollView>
-    </View>
+    <Structure title="Sign up now" subtitle="Join BrainMe and invite friends">
+      {!pendingVerification ? (
+        <>
+          <Input
+            title="Email"
+            placeholder="winner@email.com"
+            keyboardType="email-address"
+            onChangeText={setEmailAddress}
+          />
+          <Input
+            title="Your username"
+            placeholder="Rookie"
+            onChangeText={setUsername}
+          />
+          <Input
+            title="Password"
+            placeholder="Insert password..."
+            secureTextEntry
+            onChangeText={setPassword}
+          />
+          <Button text="SIGN UP" onPress={onSignUpPress} />
+          <Footer text="Already have an account?" />
+        </>
+      ) : (
+        <>
+          <Text>
+            Please, insert the verification code we provided on your email
+            address.
+          </Text>
+          <Input
+            title="Verification code"
+            placeholder="Insert code..."
+            keyboardType="numeric"
+            textAlign="center"
+            maxlength={6}
+            onChangeText={setCode}
+          />
+          <Button text="VERIFY" onPress={onPressVerify} />
+          <Footer text="Already have an account?" />
+        </>
+      )}
+    </Structure>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "white",
-    borderTopLeftRadius: 30,
-    borderTopRightRadius: 30,
-    paddingHorizontal: 17,
-    paddingTop: 51,
-  },
-});
