+++ BrainMe/components/auth/footer-text.tsx
