+++ BrainMe/package.json
@@ -21,11 +21,11 @@
     "@expo/vector-icons": "^14.0.0",
     "@react-navigation/native": "^6.0.2",
     "convex": "^1.12.1",
-    "expo": "~51.0.11",
+    "expo": "^51.0.14",
     "expo-constants": "~16.0.2",
     "expo-font": "~12.0.7",
     "expo-linking": "~6.3.1",
-    "expo-router": "~3.5.15",
+    "expo-router": "~3.5.16",
     "expo-secure-store": "~13.0.1",
     "expo-splash-screen": "~0.27.5",
     "expo-status-bar": "~1.12.1",
